package com.roxstudio: GfxUtil: 添加几个位图填充函数 AutoplaySprite: 自动播放及停止的动画精灵，适用于显示载入动画等 RoxScreenManager & RoxScreen: 支持更复杂的跳转方式，允许结束多层Screen后，跳转到一个新Screen UiUtil: 添加delay函数用来延迟执行一段代码；添加message函数用来显示一条自动消隐的消息框 package com.weiplus.client: apps.*.Maker: 生成带遮罩的缩略图 CommentsScreen: 添加评论Screen HarryCamera: 从相机Activity取消返回后不再出现空白屏幕 HaxeStub: 支持弹出本地输入框 HomeScreen, SelectedScreen, TimelineScreen: 使用纯Haxe的HpApi请求数据，解决异常崩溃问题（跟调用Android的Http请求API有关） HpApi: 对HarryPhoto REST API的包装，现在除登录和上传文件之外，其它的get请求都可用纯haxe完成，减少与Android端的交互 HpManager: 支持对第三方绑定进行开关操作，此操作不会导致重新登录绑定 LoginScreen: 纯Haxe的登录界面，支持三种第三方登录 MakerList: 更换新界面 MakerScreen: 修正魔法相机生成的照片发红的问题 MyUtils: 载入动画、时间转字符串、登出等 PostScreen: 支持对第三方绑定进行开关操作，对未绑定或绑定失效的可以直接弹出第三方登陆页面进行绑定 Postit: 支持赞与评论等 PublicScreen: 未登录的精选时间轴 RoutineScreen: 消息界面 Splash: 可在闪屏时载入一些缓存数据 UserScreen: 换用HpApi请求时间轴数据；可以根据参数查看其他用户的信息；界面更改； Binding: 和Android端对应的枚举类 model.*: 根据实际数据修改 weiplus.nmml: 更换图标 其它： clear_res.bat: 清除几个目标路径的资源，防止构建时继续使用旧资源 GameActivity.java: 为解决中文输入问题 补所有图片资源

+++ haxeone/src/com/roxstudio/haxe/io/FileUtil.hx
@@ -36,6 +36,7 @@ class FileUtil {
 
     public static function rmdir(path: String, ?force: Bool = false) {
         path = path.replace("\\", "/");
+        if (!FileSystem.exists(path) || !FileSystem.isDirectory(path)) return;
         if (!force) { FileSystem.deleteDirectory(path); return; }
         for (name in FileSystem.readDirectory(path)) {
             var sub = path + "/" + name;
